Espectogramas/requirements.txt
librosa
praat-parselmouth 
matplotlib
numpy
seaborn
pylint
pep8
autopep8
setuptools>=46.4.0
jupyterlab
pylint
ipykernel
xeus-python
notebook